Finished some rendering, made a few bugs

+++ Engine/Graphics/Direct3D12/D3D12CommonHeaders.h
@@ -37,15 +37,15 @@ if (FAILED(x)) {							\
 }
 #endif
 #define NAME_D3D12_OBJECT(obj, name) obj->SetName(name); OutputDebugString(L"::D3D12 Object Created: "); OutputDebugString(name); OutputDebugString(L"\n");
-#define NAME_D3D12_OBJECT_INDEXED(obj, idx, name)			\
-{															\
-	wchar_t full_name[128];									\
-	if (swprintf_s(full_name, L"%s[%u]", name, idx) > 0) {	\
-		obj->SetName(full_name);							\
-		OutputDebugString(L"::D3D12 Object Created: ");		\
-		OutputDebugString(full_name);						\
-		OutputDebugString(L"\n");							\
-	}														\
+#define NAME_D3D12_OBJECT_INDEXED(obj, idx, name)					\
+{																	\
+	wchar_t full_name[128];											\
+	if (swprintf_s(full_name, L"%s[%llu]", name, (u64)idx) > 0) {	\
+		obj->SetName(full_name);									\
+		OutputDebugString(L"::D3D12 Object Created: ");				\
+		OutputDebugString(full_name);								\
+		OutputDebugString(L"\n");									\
+	}																\
 }
 #ifndef DEBUG_OP
 #define DEBUG_OP(x) x
